+++ index.js
@@ -4,9 +4,9 @@ import {
 } from 'react-native'
 
 // In-house dependencies
-import App from './src/App'
+import Root from './src/Root'
 import {
     name as appName
 } from './app.json'
 
-AppRegistry.registerComponent(appName, () => App)
+AppRegistry.registerComponent(appName, () => Root)