+++ CHANGELOG.md
@@ -15,6 +15,7 @@ the code was deployed.
 - Local deployment instructions to README (CU-d8qufg).
 - DotEnv (CU-d8qufg).
 - Automatic linting and testing using Travis (CU-d8qufg).
+- Storybook (CU-d8qufg).
 
 ### Removed
 - `yarn.lock` (CU-d8qufg).